requirements.txt
pillow==12.1.1
psycopg==3.3.2
psycopg-binary==3.3.2
ttkbootstrap==1.20.1
typing_extensions==4.15.0
